stupid mistake in return value

+++ include/PolyhedralFunction.h
@@ -858,7 +858,7 @@ class PolyhedralFunction : public C05Function {
  /** Get a specific float (double) numerical parameter. PolyhedralFunction
   * takes care of dblAAccMlt, leaving all the rest to C05Function. */
 
- int get_dbl_par( const idx_type par ) const override final {
+ double get_dbl_par( const idx_type par ) const override final {
   if( par == dblAAccMlt )
    return( AAccMlt );
   else
